Add a few missing boost include file (multiple files) Cleanup include files not needed (ongoing) Replace old formid handling with RE::FormID (ongoing) Modernize hashmap eraser form_observer::u_remove_expired_forms Fix Lambda in load/save collection method parameters Fix missing form_handling (use method from forms) Fix json_ref operators and add one more variant Another case of boost variant fixes (use std::visit) Fix integer warning in ITypes header

+++ dep/common/common/ITypes.h
@@ -45,7 +45,7 @@ inline UInt64 Swap64(UInt64 in)
 {
 	UInt64	temp;
 
-	temp = Swap32(in);
+    temp = Swap32(static_cast<UInt32>(in));
     temp <<= static_cast<std::uint32_t>(32);
 	temp |= Swap32(in >> 32);
